agents/face/requirements.txt
flask
insightface>=0.7.3
numpy
pillow
